requirements.txt
Flask
torch
torchvision
Pillow
flask_sqlalchemy==3.1.1
flask_login==0.6.3
cryptography==42.0.5
bcrypt==4.1.2
